Bug fix in Enclave
Deleting a KeyPair was not possible because when the method was executed inside of "promise.resolve" it returned "Kotlin.Unit" which is not supported by react, and therefore threw an error message. Fixed it so that if there is an exception, it gets returned, if not it gets resolved with an empty string.