Add android support (#155)
* Mostly autogenerated

Mostly autogenerated code, except for AndroidManifest addition of the correct intent-filter. And in the build.gradle files the addition of the fonts.gradle file and google services for notification.

* Add firebase generated json

* Add android config in app.json

* Add config to include icons for android

* Add commands for android, and packages for android

* Load fonts for android

* Add android notification support

* Update Share to work on android

* Bug fix and change

Fixed it so that parseLocalFloat returns 0 if str is undefined.
Changed the AmountInput to use "onEndEditing" instead of "onBlur", because on android the value is not passed through like on iOS. Still works the same though.

* Bug fix in Enclave

Deleting a KeyPair was not possible because when the method was executed inside of "promise.resolve" it returned "Kotlin.Unit" which is not supported by react, and therefore threw an error message. Fixed it so that if there is an exception, it gets returned, if not it gets resolved with an empty string.

* Fix bugs

* Remove unnecessary flag

* Remove auto-generated commands

* Remove devDependency

* Fix to pull config from file

* Update header and fix null check

* fix: working icons, no gradle mods

* fix: build warnings

* fix: android build

* fix: iOS build, Android notifications

fix build warnings

* chore: new stage builds

* Fix cursor alignment bug

* Make build create apk

* Fix app link

* Rewrite share for android

* eas submit for Android

Publishes to Play Store internal test track

* Undo temp api change

* Fix linting

* Fix linting

* Add android dev quickstart

---------

Co-authored-by: DC <dc@dcpos.ch>