daimo-api: use a shovel database to load decoded logs (#485)
Replace calls to eth_getLogs with queries to a postgres database.

This is a pretty big commit. Here is a summary of the changes:

1. Introduce a Watcher object that connects to a postgres database,
finds the latest block that shovel has processed, and then compare
that number with the latest block that the Watcher has processed
and then process the difference.

2. Each indexer (coin, key, name, note, op) now exposes a load(pg,
from, to) method that the Watcher will call. These methods load
their respective data from the pg connection and then store the
decoded log data in their data structures as was happening prior
to this commit. The big change here is that instead of using eth
log related datasctructures, we are using basic objects with fields
that match the columns in the database.

Overall the data flow of the api hasn’t changed. The API still will
replays all log data on startup, and incrementallly after startup,
so that it can maintain a set of data structures that are used by
the API endpoints and push notifications.