daimo-api: bugfix. return latest block for getAccountHistory (#519)
Prior to this commit, and prior to shovel, the getAccountHistory
function would use the ViemClient to get the latest block. This assumed
that the ViemClient would be updating its block. Now that we are using
Watcher, ViemClient was no longer updating its block which means that it
latest value was only set on startup and then never updated.

Now you can ask Watcher for the latest block.