daimo-api: fix account pubkey
Prior to this commit, we were leaving the order of keys returned from
key_{added,removed} undefined. I guess it worked for a while, but in
testing I found that the getAccountHistory endpoint was returning
invalid keys. Now that the order is defined, the
contractFriendlyKeyToDER function should produce correct results.